+++ packages/react-tinacms-inline/src/styles/focus-ring.tsx
@@ -16,16 +16,23 @@ limitations under the License.
 
 */
 import styled, { css } from 'styled-components'
+import { BlocksEmptyState } from '../blocks/inline-field-blocks'
 
-interface FocusRingProps {
-  active: boolean
-  offset?: number
+export interface FocusRingStyleProps {
+  offset?: number | { x: number; y: number }
   borderRadius?: number
-  disableHover?: boolean
 }
 
-export const FocusRing = styled.div<FocusRingProps>(
-  p => css`
+interface FocusRingProps extends FocusRingStyleProps {
+  active: boolean
+  disableHover?: boolean
+  disableChildren?: boolean
+}
+
+export const FocusRing = styled.div<FocusRingProps>(p => {
+  const offset = getOffset(p.offset)
+
+  return css`
     position: relative;
     width: 100%;
 
@@ -38,15 +45,27 @@ export const FocusRing = styled.div<FocusRingProps>(
         }
       `}
 
+    ${p.disableChildren &&
+      css`
+        > * {
+          pointer-events: none;
+        }
+
+        ${BlocksEmptyState} {
+          opacity: 0;
+          pointer-events: none;
+        }
+      `}
+
     &:after {
       content: '';
       box-sizing: border-box;
       display: block;
       position: absolute;
-      left: calc(-1 * ${p.offset !== undefined ? p.offset : '16'}px);
-      top: calc(-1 * ${p.offset !== undefined ? p.offset : '16'}px);
-      width: calc(100% + ${p.offset !== undefined ? p.offset * 2 : '32'}px);
-      height: calc(100% + ${p.offset !== undefined ? p.offset * 2 : '32'}px);
+      left: calc(-1 * ${getOffsetX(offset)}px);
+      top: calc(-1 * ${getOffsetY(offset)}px);
+      width: calc(100% + ${getOffsetX(offset) * 2}px);
+      height: calc(100% + ${getOffsetY(offset) * 2}px);
       border: 1px solid var(--tina-color-primary);
       border-radius: ${p.borderRadius !== undefined ? p.borderRadius : `10`}px;
       opacity: 0;
@@ -63,4 +82,28 @@ export const FocusRing = styled.div<FocusRingProps>(
         }
       `};
   `
-)
+})
+
+export function getOffset(
+  offset: number | undefined | { x: number; y: number }
+): number | { x: number; y: number } {
+  const DEFAULT_OFFSET: number = 16
+  let result: number | { x: number; y: number } = DEFAULT_OFFSET
+  const axis = { x: DEFAULT_OFFSET, y: DEFAULT_OFFSET }
+
+  if (typeof offset === 'number') {
+    result = offset
+  } else if (typeof offset === 'object') {
+    axis.x = offset.x
+    axis.y = offset.y
+    result = axis
+  }
+
+  return result
+}
+
+export const getOffsetX = (offset: number | { x: number; y: number }): number =>
+  typeof offset === 'object' ? offset.x : offset
+
+export const getOffsetY = (offset: number | { x: number; y: number }): number =>
+  typeof offset === 'object' ? offset.y : offset