+++ packages/@tinacms/icons/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.14.0](https://github.com/tinacms/tinacms/compare/@tinacms/icons@0.14.0-alpha.0...@tinacms/icons@0.14.0) (2020-06-15)
+
+**Note:** Version bump only for package @tinacms/icons
+
+
+
+
+
+# [0.14.0-alpha.0](https://github.com/tinacms/tinacms/compare/@tinacms/icons@0.13.0...@tinacms/icons@0.14.0-alpha.0) (2020-06-12)
+
+
+### Features
+
+* add exit icon ([3c544c3](https://github.com/tinacms/tinacms/commit/3c544c3))
+* add reorder row icon ([18ba2cc](https://github.com/tinacms/tinacms/commit/18ba2cc))
+
+
+
+
+
 # [0.13.0](https://github.com/tinacms/tinacms/compare/@tinacms/icons@0.12.0...@tinacms/icons@0.13.0) (2020-06-08)