+++ packages/gatsby-tinacms-mdx/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.20](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-mdx@0.1.20-alpha.0...gatsby-tinacms-mdx@0.1.20) (2020-06-15)
+
+**Note:** Version bump only for package gatsby-tinacms-mdx
+
+
+
+
+
+## [0.1.20-alpha.0](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-mdx@0.1.19...gatsby-tinacms-mdx@0.1.20-alpha.0) (2020-06-12)
+
+**Note:** Version bump only for package gatsby-tinacms-mdx
+
+
+
+
+
 ## [0.1.19-alpha.0](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-mdx@0.1.18...gatsby-tinacms-mdx@0.1.19-alpha.0) (2020-06-08)
 ## [0.1.19](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-mdx@0.1.18...gatsby-tinacms-mdx@0.1.19) (2020-06-08)